Remove extra semicolons. Fix volatile on atomic to remove tons of warnings. Fix signed/unsigned warning.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@32191 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/utility/StringHelpers.C
@@ -1244,7 +1244,7 @@ StringHelpers::ParseRange(const string range, std::vector<int> &list)
     std::vector<std::string> rangeTokens = StringHelpers::split(range, ',');
 
     bool parseError = false;
-    for(int i =0; i < rangeTokens.size(); ++i)
+    for(size_t i =0; i < rangeTokens.size(); ++i)
     {
         int first = 0;
         int last = 0;
